# javalin-api/src/main/java/za/co/no9/ses8/adaptors/api/javalin/NewEventBean.java
package za.co.no9.ses8.adaptors.api.javalin;


public class NewEventBean {
    public String name;
    public String content;


    public NewEventBean(String name, String content) {
        this.name = name;
        this.content = content;
    }
}